tests/requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
allure-pytest==2.13.2
webdriver-manager==4.0.1
requests==2.31.0
faker==20.1.0
python-dotenv==1.0.0
pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.3
beautifulsoup4==4.12.2
